requirements.txt
lxml==4.6.4
pandas
vsts==0.1.25
pyyaml==6.0
azure-devops==6.0.0b4